frontend/requirements.txt
streamlit>=1.36.0
requests
pandas
python-dotenv
